# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kenya SMEs Organization API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OrgStats {
  totalMembers: number;
  totalCoordinators: number;
  totalSectors: number;
  countiesCovered: number;
  constituenciesCovered: number;
  wardsCovered: number;
  upcomingEvents: number;
}

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  active: 'active',
  pending: 'pending',
  suspended: 'suspended',
} as const;

export interface Member {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  businessName: string;
  businessType: string;
  category: string;
  county: string;
  constituency: string;
  ward: string;
  membershipNumber: string;
  status: MemberStatus;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  phone: string;
  /** @minLength 2 */
  businessName: string;
  businessType: string;
  category: string;
  county: string;
  constituency: string;
  ward: string;
}

export interface MemberCategory {
  id: number;
  name: string;
  description: string;
  annualFee: number;
  benefits: string[];
}

export type CoordinatorLevel = typeof CoordinatorLevel[keyof typeof CoordinatorLevel];


export const CoordinatorLevel = {
  county: 'county',
  constituency: 'constituency',
  ward: 'ward',
} as const;

export type CoordinatorStatus = typeof CoordinatorStatus[keyof typeof CoordinatorStatus];


export const CoordinatorStatus = {
  active: 'active',
  pending: 'pending',
  suspended: 'suspended',
} as const;

export interface Coordinator {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  nationalId: string;
  county: string;
  constituency: string;
  ward: string;
  level: CoordinatorLevel;
  registrationFee: number;
  status: CoordinatorStatus;
  coordinatorCode: string;
  createdAt: string;
}

export type CoordinatorInputLevel = typeof CoordinatorInputLevel[keyof typeof CoordinatorInputLevel];


export const CoordinatorInputLevel = {
  county: 'county',
  constituency: 'constituency',
  ward: 'ward',
} as const;

export interface CoordinatorInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  phone: string;
  nationalId: string;
  county: string;
  constituency: string;
  ward: string;
  level: CoordinatorInputLevel;
  mpesaNumber: string;
  motivation?: string;
}

export interface Benefit {
  id: number;
  title: string;
  description: string;
  icon: string;
}

export interface Sector {
  id: number;
  name: string;
  description: string;
  imageUrl: string;
  totalSMEs: number;
  annualContribution: string;
  keyProducts: string[];
  challenges: string[];
  opportunities: string[];
  counties: string[];
}

export type EventEventType = typeof EventEventType[keyof typeof EventEventType];


export const EventEventType = {
  workshop: 'workshop',
  conference: 'conference',
  training: 'training',
  networking: 'networking',
  exhibition: 'exhibition',
  other: 'other',
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  eventDate: string;
  /** @nullable */
  endDate?: string | null;
  venue: string;
  county: string;
  eventType: EventEventType;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  registrationDeadline?: string | null;
  /** @nullable */
  maxAttendees?: number | null;
  currentAttendees: number;
  isFeatured: boolean;
  createdAt: string;
}

export type EventInputEventType = typeof EventInputEventType[keyof typeof EventInputEventType];


export const EventInputEventType = {
  workshop: 'workshop',
  conference: 'conference',
  training: 'training',
  networking: 'networking',
  exhibition: 'exhibition',
  other: 'other',
} as const;

export interface EventInput {
  /** @minLength 3 */
  title: string;
  description: string;
  eventDate: string;
  /** @nullable */
  endDate?: string | null;
  venue: string;
  county: string;
  eventType: EventInputEventType;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  registrationDeadline?: string | null;
  /** @nullable */
  maxAttendees?: number | null;
  isFeatured?: boolean;
}

export interface EventRegistrant {
  id: number;
  eventId: number;
  fullName: string;
  email: string;
  phone: string;
  createdAt: string;
}

export interface EventRegistrantInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  phone: string;
}

export type ListMembersParams = {
category?: string;
county?: string;
limit?: number;
offset?: number;
};

export type ListCoordinatorsParams = {
county?: string;
constituency?: string;
limit?: number;
offset?: number;
};

export type ListEventsParams = {
upcoming?: boolean;
limit?: number;
};