# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kenya SMEs Organization API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: county
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: registerMember
      tags: [members]
      summary: Register a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Validation error
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        category:
          type: string
        county:
          type: string
        constituency:
          type: string
        ward:
          type: string
        membershipNumber:
          type: string
        status:
          type: string
          enum: [active, pending, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - phone
        - businessName
        - businessType
        - category
        - county
        - constituency
        - ward
        - membershipNumber
        - status
        - createdAt
    MemberInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        businessName:
          type: string
          minLength: 2
        businessType:
          type: string
        category:
          type: string
        county:
          type: string
        constituency:
          type: string
        ward:
          type: string
      required:
        - fullName
        - email
        - phone
        - businessName
        - businessType
        - category
        - county
        - constituency
        - ward
...